make base node and edge attributes no longer forward declared
