+++ include/spark_dsg/serialization/graph_json_serialization.h
@@ -33,6 +33,7 @@
  * purposes notwithstanding any copyright notation herein.
  * -------------------------------------------------------------------------- */
 #pragma once
+#include <memory>
 #include <string>
 
 #include "spark_dsg/spark_dsg_fwd.h"
