+++ include/spark_dsg/spark_dsg_fwd.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <memory>
 
 namespace spark_dsg {
